cambios

+++ src/services/server_services/server_service.js
@@ -27,15 +27,19 @@ const find_all_enabled_services_by_name = async(name) => {
 const deleteServiceByIds = async(id) => {
     const body = {id: id}
     const recordToRemove = await find_all_enabled_services_by_id(id)
-    const {url, authorization} = await godaddyUrlAndAuthorization()
-    const arr = [recordToRemove.type, recordToRemove.service_name]
-    const buildedUrl = utils.buildGodaddyUrl(url, arr)
-    const godaddyResponse = await utils.delete_external_api_with_security(buildedUrl, authorization)
-    if(godaddyResponse === 204){
-        const response = await query(query_utils.delete_enabled_services_by_ids(), body)
-        return response
+    if(Object.keys(recordToRemove).length) {
+        const {url, authorization} = await godaddyUrlAndAuthorization()
+        const arr = [recordToRemove.type, recordToRemove.service_name]
+        const buildedUrl = utils.buildGodaddyUrl(url, arr)
+        const godaddyResponse = await utils.delete_external_api_with_security(buildedUrl, authorization)
+        if(godaddyResponse === 204){
+            const response = await query(query_utils.delete_enabled_services_by_ids(), body)
+            return response
+        }else {
+            return godaddyResponse
+        }
     }else {
-        return godaddyResponse
+        return {status: 404, message: "field not found"}
     }
 
 }
@@ -46,7 +50,8 @@ const saveNewEnabledService = async (body) => {
     const goDaddyBody = [{ data: public_ip, name: body.service_name, ttl: 600, type: "A" }]
     const godaddyResponse = await utils.patch_external_api_with_security(url, goDaddyBody, authorization);
     if (godaddyResponse === 200) {
-        const response = await query(query_utils.save_new_enabled_service, body)
+        const bodyQuery = {service_name: body.service_name, ttl: 600, type: "A"}
+        const response = await query(query_utils.save_new_enabled_service, bodyQuery)
         return {
             status: 201,
             message: 'field suscessfully created',
